requirements.txt
requests>=2.31.0
urllib3>=2.0.0
python-dotenv>=1.0.0
tenacity>=8.2.0
python-telegram-bot>=20.0
